Time: 8 ms (76.95%) | Memory: 42.9 MB (92.21%) - LeetSync

+++ 2562-count-ways-to-build-good-strings/count-ways-to-build-good-strings.java
@@ -3,10 +3,10 @@ class Solution {
     public int mod = 1000000000+7;
     public int countGoodStrings(int low, int high, int zero, int one) {
         
-        long[] memo = new long[high+1];
+        int[] memo = new int[high+1];
         memo[zero]++;
         memo[one]++;
-        long ans =0;
+        int ans =0;
         int start = (zero<one)?zero : one;
         for(int i = start;i<=high;i++){
             if(i>=low) ans = (ans+memo[i])%mod;
